+++ mycli/packages/sqlresult.py
@@ -1,5 +1,7 @@
 from dataclasses import dataclass
+from functools import cached_property
 
+from prompt_toolkit.formatted_text import FormattedText, to_plain_text
 from pymysql.cursors import Cursor
 
 
@@ -9,7 +11,7 @@ class SQLResult:
     header: list[str] | str | None = None
     rows: Cursor | list[tuple] | None = None
     postamble: str | None = None
-    status: str | None = None
+    status: str | FormattedText | None = None
     command: dict[str, str | float] | None = None
 
     def __iter__(self):
@@ -17,3 +19,9 @@ class SQLResult:
 
     def __str__(self):
         return f"{self.preamble}, {self.header}, {self.rows}, {self.postamble}, {self.status}, {self.command}"
+
+    @cached_property
+    def status_plain(self):
+        if self.status is None:
+            return None
+        return to_plain_text(self.status)